// generated/api/apiSchemas.ts
export type InternalID = string | number;
// ...
export type SignatureAlgorithm =
  | "HS256"
  | "HS384"
  | "HS512"
  | "RS256"
  | "RS384"
  | "RS512"
  | "ES256"
  | "ES384"
  | "ES512"
  | "PS256"
  | "PS384"
  | "PS512";
// ...
export type Jwk = {
  /**
   * The key type (e.g., EC).
   *
   * @example EC
   */
  kty: string;
  /**
   * The elliptic curve used (only for EC keys).
   *
   * @example P-256
   */
  crv?: string | null;
  /**
   * The key ID.
   *
   * @example 12345
   */
  kid: string;
  /**
   * The X coordinate for EC keys (optional).
   *
   * @example o-7zraXKDaoBte2PsuTXo-MSLzsyWdAElNptGgI4aH8
   */
  x?: string | null;
  /**
   * The Y coordinate for EC keys (optional).
   *
   * @example Xr_wCzJ1XnsgAIV5qHruzSwaNnwy87UjmevVklTpIv8
   */
  y?: string | null;
  /**
   * The modulus for RSA keys.
   *
   * @example modulus_value
   */
  n?: string | null;
  /**
   * The exponent for RSA keys.
   *
   * @example AQAB
   */
  e?: string | null;
  alg?: SignatureAlgorithm;
  /**
   * The intended use of the key (e.g., sig, enc).
   *
   * @example sig
   */
  use?: string | null;
  /**
   * A URL that points to an X.509 public key certificate or certificate chain.
   *
   * @format uri
   * @example https://example.com/cert.pem
   */
  x5u?: string | null;
  /**
   * The X.509 certificate chain.
   */
  x5c?: string[] | null;
  /**
   * The SHA-1 thumbprint of the X.509 certificate.
   *
   * @example dGhpcyBpcyBqdXN0IGEgdGh1bWJwcmludA
   */
  x5t?: string | null;
  /**
   * The SHA-256 thumbprint of the X.509 certificate.
   *
   * @example sM4KhEI1Y2Sb6-EVr6tJabmJuoP-ZE...
   */
  x5tS256?: string | null;
};

/**
 * @example {"alg":"ES512","crv":"P-521","kid":"LxIDw8zAmgI2OjsXMS_QKyiPaR7GH1QVw3sIaUlWWy0","kty":"EC","use":"sig","x":"AJgCXH_ugrh2AgOgNaD2h4ho36fobGDvUn5BzWOGAEF--sfRW6bT5gHx_kNh6ajnVCbPdOn3kfx6fCX306MmfmNb","y":"AON1eYFPmxeFkNYD04Ar6rToYruV8_btso-iXAkLRpOBrLeICdjxUuyPnFpOHmRbyoqih7tla2Nwo1dle8JJKEC8"}
 */
export type ManagedJWK = Jwk & {
  /**
   * The Key Management System
   *
   * @example filesystem
   */
  kms?: string;
  /**
   * The KMS Key reference (the id of the key in the KMS) if supported
   *
   * @example my-key
   */
  kms_key_ref?: string;
  /**
   * The timestamp when the JWK was created.
   *
   * @example 1756449710
   */
  iat?: number | null;
  /**
   * The time when the key expires and will not be valid anymore.
   *
   * @example 1756536110
   */
  exp?: number;
};
// ...
export type TrustMark = {
  id: InternalID;
  /**
   * The trust mark type identifier.
   */
  trust_mark_type: string;
  /**
   * The issuer of the trust mark.
   */
  trust_mark_issuer: string;
  /**
   * The trust mark JWT.
   */
  trust_mark: string;
};